pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "tucan"
authors = [{name = "Jan C. Brammer", email = "jan.c.brammer@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
dynamic = ["version", "description"]

dependencies = [
    "networkx[default] >=2.7.1",
    "matplotlib >=3.5.0",
    "tabulate >=0.8.9",
    "igraph >=0.9.10",
    "antlr4-python3-runtime == 4.11.1"
]

[project.optional-dependencies]
drawing = [
    "plotly",    # interactive 3D drawings
    "notebook",
    "ipywidgets"
]
dev = [
    "black",
    "flake8",
    "pytest"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = """
    -v
    --durations=5"""
testpaths = [
    "tests",
]
